+++ NEWS.md
@@ -1,11 +1,22 @@
-# 0.2.1: 2020-04-29
+# modelsummary 0.3.0
+
+* Out-of-the-box Rmarkdown compilation to HTML, PDF, RTF
+* kableExtra output format for LaTeX and Markdown
+* Support for `threeparttable`, colors, and many other LaTeX options
+* Deprecated arguments: filename, subtitle
+* Deprecated functions: clean_latex, knit_latex
+* `pkgdown` website and doc improvements
+* `mitools` tidiers
+* New tests
+
+# modelsummary 0.2.1
 
 * Convenience function to render markdown in row/column labels
 * bug: breakage when all GOF were omitted
 * Clean up manual with @keywords internal
 * bug: tidyr import
 
-# 0.2.0: 2020-04-04
+# modelsummary 0.2.0
 
 * gt is now available on CRAN
 * new latex_env argument for knit_latex and clean_latex
@@ -21,6 +32,6 @@
 * Internal functions have @keywords internal to avoid inclusion in docs
 * Statistic override accepts pre-formated character vectors
 
-# 0.1.0: 2019-06-30
+# modelsummary 0.1.0
 
 * Initial release (gt still needs to be installed from github)